images/tima_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
Pillow>=10.0.0,<11.0
django-filter>=23.3,<24.0
coverage>=7.2.0,<8.0
pytest>=7.4.0,<8.0
pytest-django>=4.5.2,<5.0
python-dotenv>=1.0.0,<2.0
python-decouple>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
drf-yasg>=1.21,<2.0
django-rest-passwordreset>=1.3,<2.0
django-allauth>=0.61,<1.0
django-allauth-ui>=0.2.0,<0.3.0
requests